% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tibble_row}
\alias{as_tibble_row}
\title{Convert list to tibble row}
\usage{
as_tibble_row(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A tibble with 1 row.
}
\description{
Convert list to tibble row
}
\keyword{internal}
