% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_string}
\alias{check_string}
\title{Validation helpers}
\usage{
check_string(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check.}

\item{arg_name}{Name of the argument for error messages.}
}
\description{
Validation helpers
}
\keyword{internal}
