% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{create_session}
\alias{create_session}
\title{Create Session}
\usage{
create_session(session_data = list())
}
\arguments{
\item{session_data}{List containing session configuration}
}
\value{
The created session
}
\description{
Create Session
}
