% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{items_to_tibble}
\alias{items_to_tibble}
\title{Convert list of items to tibble}
\usage{
items_to_tibble(items)
}
\arguments{
\item{items}{List of lists.}
}
\value{
A tibble.
}
\description{
Convert list of items to tibble
}
\keyword{internal}
