% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{jules_auth}
\alias{jules_auth}
\title{Authenticate with Jules}
\usage{
jules_auth(api_key = NULL)
}
\arguments{
\item{api_key}{Jules API key. If \code{NULL}, it looks for the \code{JULES_API_KEY}
environment variable.}
}
\value{
Invisibly returns the API key.
}
\description{
Authenticate with Jules
}
\examples{
\dontrun{
jules_auth("your_api_key")
}
}
