% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{jules_perform}
\alias{jules_perform}
\title{Perform Jules API request}
\usage{
jules_perform(req)
}
\arguments{
\item{req}{httr2 request}
}
\value{
Parsed JSON response
}
\description{
Perform Jules API request
}
\keyword{internal}
