% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{jules_perform_all}
\alias{jules_perform_all}
\title{Perform paginated Jules API request}
\usage{
jules_perform_all(path, key, query = list())
}
\arguments{
\item{path}{API path}

\item{key}{Key for the list in the response (e.g., "sources")}

\item{query}{Optional query parameters}
}
\value{
A list of all items
}
\description{
Perform paginated Jules API request
}
\keyword{internal}
