% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{jules_request}
\alias{jules_request}
\title{Base request for Jules API}
\usage{
jules_request(path)
}
\arguments{
\item{path}{API path}
}
\value{
A httr2 request object
}
\description{
Base request for Jules API
}
\keyword{internal}
