% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{jules_sessions}
\alias{jules_sessions}
\title{List Sessions}
\usage{
jules_sessions()
}
\value{
A tibble of sessions.
}
\description{
List Sessions
}
\examples{
if (nzchar(Sys.getenv("JULES_API_KEY"))) {
  jules_sessions()
}
}
