% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{jules_sources}
\alias{jules_sources}
\title{List Sources}
\usage{
jules_sources()
}
\value{
A tibble of sources.
}
\description{
List Sources
}
\examples{
if (nzchar(Sys.getenv("JULES_API_KEY"))) {
  jules_sources()
}
}
