% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{list_activities}
\alias{list_activities}
\title{List Activities}
\usage{
list_activities(parent)
}
\arguments{
\item{parent}{Parent session name (e.g., "sessions/123")}
}
\value{
A tibble of activities
}
\description{
List Activities
}
\examples{
if (nzchar(Sys.getenv("JULES_API_KEY"))) {
  list_activities("sessions/my-session")
}
}
