% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{list_sessions}
\alias{list_sessions}
\title{List Sessions}
\usage{
list_sessions()
}
\value{
A tibble of sessions
}
\description{
List Sessions
}
\examples{
if (nzchar(Sys.getenv("JULES_API_KEY"))) {
  list_sessions()
}
}
