% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{list_sources}
\alias{list_sources}
\title{List Sources}
\usage{
list_sources()
}
\value{
A tibble of sources
}
\description{
List Sources
}
\examples{
if (nzchar(Sys.getenv("JULES_API_KEY"))) {
  list_sources()
}
}
