% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jules_design.R
\name{write_jules_design}
\alias{write_jules_design}
\title{Create an ensemble design for jules}
\usage{
write_jules_design(
  X_mm = NULL,
  paramlist,
  n,
  fac,
  minfac,
  maxfac,
  tf,
  fnprefix = "param-perturb-test",
  lhsfn = "lhs.txt",
  stanfn = "stanparms.txt",
  allstanfn = "allstanparms.txt",
  rn = 5,
  startnum = 0
)
}
\arguments{
\item{X_mm}{augnmented design from addNroyDesignPoint}

\item{paramlist}{character vector of chosen parameters}

\item{n}{number of ensemble members}

\item{fac}{character vector of names of variables that you would like to alter by a factor. Everything else gets variaed by PFT}

\item{minfac}{must correspond to fac - i.e. one value per parameter,}

\item{maxfac}{must correspond to fac - i.e. one value per parameter,}

\item{tf}{character vector containing the logical parameters}

\item{fnprefix}{filename prefix}

\item{lhsfn}{latin hypercube filename}

\item{stanfn}{standard parameter filename}

\item{allstanfn}{allstanparams filename}

\item{rn}{rounding number}

\item{startnum}{ensemble indexing start number}
}
\value{
none
}
\description{
This code writes a design taking either a 'factor', min and max by which to multiply all pfts,
or perturbing each pft individually according to their maximum and minimum in the parameter list.
}
