% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{send_message}
\alias{send_message}
\title{Send Message}
\usage{
send_message(name, text)
}
\arguments{
\item{name}{Session name}

\item{text}{Message text}
}
\value{
Response from the API
}
\description{
Send Message
}
